+++ flatlaf-core/src/main/java/com/formdev/flatlaf/FlatClientProperties.java
@@ -43,13 +43,13 @@ public interface FlatClientProperties
 	String BUTTON_TYPE_SQUARE = "square";
 
 	/**
-	 * Paint the toggle button in underline style.
+	 * Paint the toggle button in tab style.
 	 * <p>
 	 * <strong>Components</strong> {@link javax.swing.JToggleButton}
 	 *
 	 * @see #TOGGLE_BUTTON_TYPE
 	 */
-	String BUTTON_TYPE_UNDERLINE = "underline";
+	String BUTTON_TYPE_TAB = "tab";
 
 	/**
 	 * Paint a help button (circle with question mark).